splitted AppController into different classes
