+++ .swiftlint.yml
@@ -1,6 +1,7 @@
+disabled_rules: # rule identifiers to exclude from running
+  - identifier_name
+  - function_parameter_count
+
 excluded: # paths to ignore during linting. Takes precedence over `included`.
   - Carthage
   - Pods
-disabled_rules: # rule identifiers to exclude from running
-  - identifier_name
-  - file_length
