+++ DiveLane/App/AppDelegate.swift
@@ -18,7 +18,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var controller: AppController!
     let transactionsService: ITransactionsService = TransactionsService()
 
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
+    func application(_ application: UIApplication,
+                     didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         Fabric.with([Crashlytics.self])
 
         window = UIWindow(frame: UIScreen.main.bounds)
@@ -28,7 +29,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         return true
     }
 
-    func application(_ app: UIApplication, open url: URL, options: [UIApplicationOpenURLOptionsKey: Any] = [:]) -> Bool {
+    func application(_ app: UIApplication,
+                     open url: URL,
+                     options: [UIApplicationOpenURLOptionsKey: Any] = [:]) -> Bool {
         controller = AppController(window: window!, launchOptions: nil, url: url)
         return true
     }
