+++ DiveLane/Presentation Layer/Views/Alerts.swift
@@ -8,35 +8,35 @@
 
 import UIKit
 
-func showErrorAlert(for viewController: UIViewController, error: Error?, completion: @escaping () -> ()) {
+func showErrorAlert(for viewController: UIViewController, error: Error?, completion: (() -> Void)?) {
     var text: String?
     if let error = error as? TransactionErrors {
         text = error.rawValue
     }
     let alert = UIAlertController(title: "Error", message: text ?? error?.localizedDescription, preferredStyle: .alert)
-    let cancelAction = UIAlertAction(title: "Cancel", style: .cancel) { (action) in
-        completion()
+    let cancelAction = UIAlertAction(title: "Cancel", style: .cancel) { (_) in
+        completion?()
     }
     alert.addAction(cancelAction)
     viewController.present(alert, animated: true, completion: nil)
 }
 
-func showSuccessAlert(for viewController: UIViewController, completion: @escaping () -> ()) {
+func showSuccessAlert(for viewController: UIViewController, completion: (() -> Void)?) {
     let alert = UIAlertController(title: "Success", message: nil, preferredStyle: .alert)
-    let cancelAction = UIAlertAction(title: "OK", style: .cancel) { (action) in
-        completion()
+    let cancelAction = UIAlertAction(title: "OK", style: .cancel) { (_) in
+        completion?()
     }
     alert.addAction(cancelAction)
     viewController.present(alert, animated: true, completion: nil)
 }
 
-func showAccessAlert(for viewController: UIViewController, with text: String?, completion: @escaping (Bool) -> ()) {
+func showAccessAlert(for viewController: UIViewController, with text: String?, completion: ((Bool) -> Void)?) {
     let alert = UIAlertController(title: text ?? "Yes?", message: nil, preferredStyle: .alert)
-    let okAction = UIAlertAction(title: "Yes", style: .default) { (action) in
-        completion(true)
+    let okAction = UIAlertAction(title: "Yes", style: .default) { (_) in
+        completion?(true)
     }
-    let cancelAction = UIAlertAction(title: "No", style: .cancel) { (action) in
-        completion(false)
+    let cancelAction = UIAlertAction(title: "No", style: .cancel) { (_) in
+        completion?(false)
     }
     alert.addAction(okAction)
     alert.addAction(cancelAction)
